+++ src/PSRule.Rules.Azure/rules/Azure.Resource.Rule.ps1
@@ -8,6 +8,7 @@ if ($Null -eq $Configuration.azureAllowedRegions) {
 
 # Synopsis: Resources should be tagged
 Rule 'Azure.Resource.UseTags' -If { (SupportsTags) } -Tag @{ severity = 'Awareness'; category = 'Operations management' } {
+    Reason $LocalizedData.ResourceNotTagged
     # List of resource that support tags can be found here: https://docs.microsoft.com/en-us/azure/azure-resource-manager/tag-support
     (Exists 'Tags') -and
     (($TargetObject.Tags.PSObject.Members | Where-Object { $_.MemberType -eq 'NoteProperty' }) -ne $Null)